requirements.txt
streamlit
langchain-huggingface
langchain-community
python-dotenv
google-generativeai
PyPDF2
deep-translator
pdfminer.six